Extract business logic
